Change name of component, add in `bootstrap()`
Seems like the name of the component is XyzComponent and that maps to an
html element <xyz></xyz> rather than <xyz-component></xyz-component>.
So I added `Name` to `MyComponent` to make it `MyNameComponent` and
changed the tag to be <my-name></my-name>.

In the latest examples I've seen the @View annotations are using
`template` rather than `inline`. I could be wrong there.

Added in a snippet with the `bootstrap()` call as that'll help people
get the example a bit closer to something that works.